requirements.txt
miniflux
ebooklib
python-dotenv
requests
beautifulsoup4
